refactor(tests): Split listTasks and clearSubtasks tests
